+++ changelog.md
@@ -5,6 +5,7 @@
 
 ### 28 August 2025:
 * Update FirelySDK dependency to 5.12.2
+* Include support for $index return type checking
 
 ### 25 July 2025:
 * Update the scope of processing variables to be $this rather than the functions input collection.